Fix Short error
change to ints

+++ src/main/java/LongSet.java
@@ -2,10 +2,10 @@ import java.util.*;
 
 public class LongSet {
   public static void main (String[] args) {
-    Set<Short> s = new HashSet<Short>();
-    for (short i = 0; i < 1000; i++) {
-      s.add(i);
-	  s.remove(i - 1);
+    Set<Integer> s = new HashSet<Integer>();
+    for (int i = 0; i < 1000; i++) {
+        s.add(i);
+	s.remove(i - 1);
     }
     System.out.println(s.size());
   }
